Fix bandwidth utilization calculation issues (Phases 1-4)
This commit implements critical fixes to the SNMP bandwidth utilization
system, addressing fundamental mathematical errors, data type issues, and
OID limitations identified in the technical audit.

## Changes Made

### Phase 1: Database Schema Migration
- Changed InterfaceMetric counter columns from Float to BigInteger
  - Eliminates precision loss on large 64-bit counters
  - Files: backend/app/core/models.py, backend/app/core/schemas.py
- Created migration script: backend/migrate_counters_to_bigint.py
  - Safely migrates existing SQLite database with automatic backup

### Phase 2A: 64-bit High-Capacity Counter Support
- Added ifHCInOctets and ifHCOutOctets OIDs (64-bit counters)
  - File: backend/app/core/schemas.py
- Implemented HC counter polling with 32-bit fallback
  - File: backend/services/polling_service.py
- Prevents counter wrap on high-speed (>1Gbps) interfaces

### Phase 2B: ifHighSpeed Support
- Added ifHighSpeed OID support (handles >4Gbps interfaces)
  - File: backend/app/core/schemas.py
- Updated calculate_interface_speed() with ifHighSpeed priority
  - File: backend/services/polling_service.py
- Correctly reports speeds for 10G/40G/100G interfaces

### Phase 3: Fix Rate Calculation Formula (CRITICAL)
- Fixed mathematically incorrect bandwidth rate formula
  - Changed: avg(delta_seconds) → sum(delta_seconds)
  - File: backend/app/api/v1/endpoints/query.py
  - Affected endpoints: /network-throughput, /device-utilization,
    /report/network-throughput
- Ensures accurate bps calculations: bps = sum(bytes) * 8 / sum(seconds)

### Phase 4: Fix Capacity Calculation
- Excluded down/disabled interfaces from capacity calculation
  - Added filters: admin_status == 1, oper_status == 1
  - File: backend/app/api/v1/endpoints/query.py
- Produces accurate utilization percentages

## Documentation
- Added comprehensive fix summary: BANDWIDTH_FIX_SUMMARY.md
  - Includes deployment instructions
  - Verification checklist
  - Rollback procedures

## Impact
✅ Eliminates precision loss on large counters
✅ Prevents counter wrap on high-speed interfaces
✅ Accurate speed reporting for 10G+ interfaces
✅ Mathematically correct bandwidth rates
✅ Realistic utilization percentages

## Testing
✅ All modified files pass Python syntax checks
✅ Backwards compatible with fallback logic
✅ No breaking changes to API

Addresses technical audit findings for bandwidth utilization calculation.

+++ backend/app/core/models.py
@@ -115,13 +115,13 @@ class InterfaceMetric(Base):
     admin_status: Mapped[int | None] = mapped_column(Integer)
     oper_status: Mapped[int | None] = mapped_column(Integer)
     
-    # Counters
-    octets_in: Mapped[float | None] = mapped_column(Float)
-    octets_out: Mapped[float | None] = mapped_column(Float)
-    errors_in: Mapped[float | None] = mapped_column(Float)
-    errors_out: Mapped[float | None] = mapped_column(Float)
-    discards_in: Mapped[float | None] = mapped_column(Float)
-    discards_out: Mapped[float | None] = mapped_column(Float)
+    # Counters (BigInteger prevents precision loss on large 64-bit counters)
+    octets_in: Mapped[int | None] = mapped_column(BigInteger)
+    octets_out: Mapped[int | None] = mapped_column(BigInteger)
+    errors_in: Mapped[int | None] = mapped_column(BigInteger)
+    errors_out: Mapped[int | None] = mapped_column(BigInteger)
+    discards_in: Mapped[int | None] = mapped_column(BigInteger)
+    discards_out: Mapped[int | None] = mapped_column(BigInteger)
     
     interface: Mapped["Interface"] = relationship(back_populates="metrics")
